+++ app/controllers/settings/base_controller.rb
@@ -1,5 +1,7 @@
 module Settings
   class BaseController < ApplicationController
+    before_action :require_signed_in_user
+
     layout "application"
   end
 end