+++ test/services/things/label_png_test.rb
@@ -15,6 +15,11 @@ class Things::LabelPngTest < ActiveSupport::TestCase
 
     assert_equal label.page_width_mm.round(1), width_mm
     assert_equal label.page_height_mm.round(1), height_mm
+
+    left_third_dark_pixels = (0...(image.width / 3)).sum do |x|
+      (0...image.height).count { |y| ChunkyPNG::Color.r(image[x, y]) < 200 }
+    end
+    assert_operator left_third_dark_pixels, :>, 100, "expected QR code pixels on the left side of the PNG label"
   ensure
     label&.cleanup!
   end